Switch entry/transaction naming in LocalLedgerDB
